# src/main/java/com/example/crudelasticsearch/model/Product.java
package com.example.crudelasticsearch.model;

import java.util.List;

import org.hibernate.validator.constraints.Range;
import org.springframework.data.annotation.Id;
import org.springframework.data.elasticsearch.annotations.Document;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
@Data
@NoArgsConstructor
@AllArgsConstructor
@Document(indexName = "products")
public class Product {

    @Id
    private Long productID;
    @Range(min=0,max=10)
    private int rating;
    private List<Review> reviews;

}