# src/main/resources/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Review","description":"Review API"},{"name":"Product","description":"Product API"}],"paths":{"/products/{productID}/reviews/{reviewID}":{"get":{"tags":["Review"],"summary":"Fetch Review","description":"Fetch a perticular review with reviewID.","operationId":"getReview","parameters":[{"name":"productID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"reviewID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"put":{"tags":["Review"],"summary":"Update review","description":"Update the review for a product","operationId":"updatReview","parameters":[{"name":"productID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"reviewID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}},"required":true},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"delete":{"tags":["Review"],"summary":"Delete Review","description":"Delete the Review.","operationId":"deletReview","parameters":[{"name":"productID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"reviewID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/products/{id}":{"get":{"tags":["Product"],"summary":"Fetch Product","description":"Fetch a perticular product with productID.","operationId":"geProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"put":{"tags":["Product"],"summary":"Update Product","description":"Update the product","operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"delete":{"tags":["Product"],"summary":"Delete Product","description":"Delete the product.","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/products":{"get":{"tags":["Product"],"summary":"Fetch Products","description":"Fetch all the products.","operationId":"getAllProducts","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":3}},{"name":"field","in":"query","required":false,"schema":{"type":"string","default":"empty"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"post":{"tags":["Product"],"summary":"Create Product","description":"Create a product by giving productID,rating and review.","operationId":"saveProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"201":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/products/{productID}/reviews":{"get":{"tags":["Review"],"summary":"Fetch Reviews","description":"Fetch all the Reviews.","operationId":"getReviews","parameters":[{"name":"productID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}}}}},"post":{"tags":["Review"],"summary":"Create Review","description":"Create a review by giving reviewID,name and description.","operationId":"creatReview","parameters":[{"name":"productID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}},"required":true},"responses":{"201":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}}},"components":{"schemas":{"Review":{"type":"object","properties":{"reviewID":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"Product":{"type":"object","properties":{"productID":{"type":"integer","format":"int64"},"rating":{"type":"integer","format":"int32"},"reviews":{"type":"array","items":{"$ref":"#/components/schemas/Review"}}}}}}}